Introduce DataHub Concept
DataHub is responsible for managing data. It modifies data when receiving actions and returns the modified state.

We also need to handle SingleLiveEvents which are represented by Announcements. These represent short lived events like toasts, or alerts which is why we use a BehaviourRelay.

Remove data management logic out of MainActivity and hook up the flow of data between View and DataHub using data transformers.

Make DynamicAdapter a Consumer of DiffUIModels so that it can be better integrated into RX